+++ src/snowflake/snowpark/modin/plugin/compiler/snowflake_query_compiler.py
@@ -138,11 +138,13 @@ from snowflake.snowpark.modin.plugin._internal.aggregation_utils import (
     column_quantile,
     convert_agg_func_arg_to_col_agg_func_map,
     drop_non_numeric_data_columns,
+    format_kwargs_for_error_message,
     generate_column_agg_info,
     generate_rowwise_aggregation_function,
     get_agg_func_to_col_map,
     get_pandas_aggr_func_name,
     get_snowflake_agg_func,
+    using_named_aggregations_for_func,
 )
 from snowflake.snowpark.modin.plugin._internal.apply_utils import (
     APPLY_LABEL_COLUMN_QUOTED_IDENTIFIER,
@@ -688,7 +690,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         if args and kwargs:
             args_str += ", "
         if kwargs:
-            args_str += ", ".join(f"{key}={value}" for key, value in kwargs.items())
+            args_str += format_kwargs_for_error_message(kwargs)
         ErrorMessage.not_implemented(
             f"Snowpark pandas doesn't yet support the method {object_type}.{fn_name}({args_str})"
         )
@@ -2657,7 +2659,6 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         Returns:
             SnowflakeQueryCompiler: with a newly constructed internal dataframe
         """
-
         level = groupby_kwargs.get("level", None)
         is_supported = check_is_groupby_supported_by_snowflake(
             by, level, axis
@@ -2676,14 +2677,11 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 )
             else:
                 # Named aggregates are passed in via agg_kwargs. We should not pass `agg_func` since we have modified
-                # it to be of the form {column_name: (agg_func, new_column_name), ...}, which will cause pandas to error out.
-                if isinstance(agg_func, dict) and all(
-                    is_named_tuple(func) and len(func) == 2
-                    for func in agg_func.values()
-                ):
-                    agg_func = ", ".join(
-                        [f"{key}={value}" for key, value in agg_kwargs.items()]
-                    )
+                # it to be of the form {column_name: (agg_func, new_column_name), ...}, which will cause the error message
+                # to be misformatted. Instead, we re-format into the form f"agg(**named_aggregations)" so that the error message
+                # is recognizable to the user.
+                if using_named_aggregations_for_func(agg_func):
+                    agg_func = format_kwargs_for_error_message(agg_kwargs)
                     agg_func = f"agg({agg_func})"
 
                 ErrorMessage.not_implemented(
@@ -2768,6 +2766,34 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         new_data_column_quoted_identifier = [
             col_agg_op.agg_snowflake_quoted_identifier for col_agg_op in agg_col_ops
         ]
+
+        # The ordering of the named aggregations is changed by us when we process
+        # the agg_kwargs into the func dict (named aggregations on the same
+        # column are moved to be contiguous, see groupby.py::aggregate for an
+        # example). We need to check if the order of the output columns is correct,
+        # and if not, reorder them.
+        if uses_named_aggs:
+            correct_ordering = list(agg_kwargs.keys())
+            if correct_ordering != new_data_column_pandas_labels:
+                # In this case, we need to reorder the new_data_column_pandas_labels
+                # and the new_data_column_quoted_identifier.
+                data_column_label_to_quoted_identifier = list(
+                    zip(
+                        new_data_column_pandas_labels, new_data_column_quoted_identifier
+                    )
+                )
+                new_data_column_pandas_labels, new_data_column_quoted_identifier = list(
+                    zip(
+                        *[
+                            pair
+                            for column_label in correct_ordering
+                            for pair in filter(
+                                lambda pair: pair[0] == column_label,
+                                data_column_label_to_quoted_identifier,
+                            )
+                        ]
+                    )
+                )
         if sort:
             # when sort is True, the result is ordered by the groupby keys
             ordering_columns = [
@@ -4237,12 +4263,23 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         numeric_only = kwargs.get("numeric_only", False)
         # Call fallback if the aggregation function passed in the arg is currently not supported
         # by snowflake engine.
+        # If we are using Named Aggregations, we need to do our supported check slightly differently.
+        uses_named_aggs = using_named_aggregations_for_func(func)
         if not check_is_aggregation_supported_in_snowflake(func, kwargs, axis):
-            ErrorMessage.not_implemented(
-                f"Aggregate function {func} with parameters "
-                + ", ".join([f"{k}={v}" for k, v in kwargs.items()])
-                + " not supported yet in Snowpark pandas."
-            )
+            if uses_named_aggs:
+                # Format the error message a little differently, so it looks nicer.
+                error_msg = (
+                    "Aggregate with func=None and parameters "
+                    + format_kwargs_for_error_message(kwargs)
+                    + " not supported yet in Snowpark pandas."
+                )
+            else:
+                error_msg = (
+                    f"Aggregate function {func} with parameters "
+                    + format_kwargs_for_error_message(kwargs)
+                    + " not supported yet in Snowpark pandas."
+                )
+            ErrorMessage.not_implemented(error_msg)
 
         query_compiler = self
         if numeric_only:
@@ -4272,6 +4309,9 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         #    max  min
         # 1  3.0  NaN
         # 0  NaN  0.0
+        # should_squeeze cannot be True if we are using named aggregations, since
+        # the values for func in that case are either NamedTuples (AggFuncWithLabels) or
+        # lists of NamedTuples, both of which are list like.
         should_squeeze = is_dict_like(func) and all(
             not is_list_like(value) for value in func.values()
         )
@@ -4352,27 +4392,6 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 func,
                 pandas_labels_for_columns_to_exclude_when_agg_on_all=[],
             )
-
-            # get a map between each aggregation function and the columns needs to apply this aggregation function
-            agg_func_to_col_map = get_agg_func_to_col_map(column_to_agg_func)
-
-            # aggregation creates an index column with the aggregation function names as its values
-            # For example: with following dataframe
-            #       A   B   C
-            #   0   1   2   3
-            #   1   4   5   6
-            #   2   7   8   9
-            # after we call df.aggregate({"A": ["min"], "B": ["max"]}), the result is following
-            #       A   B
-            # min   1   NaN
-            # max   NaN	8
-            #
-            # However, if all values in the agg_func dict are scalar strings/functions rather than lists,
-            # then the result will instead be a Series:
-            # >>> df.aggregate({"A": "min", "B": "max"})
-            # 0    1
-            # 1    8
-            # dtype: int64
 
             # generate the quoted identifier for the aggregation function name column
             agg_name_col_quoted_identifier = (
@@ -4511,19 +4530,144 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 )
 
             else:
-                for agg_func, cols in agg_func_to_col_map.items():
-                    col_single_agg_func_map = {
-                        column: AggFuncInfo(
-                            func=agg_func if column in cols else "min",
-                            is_dummy_agg=column not in cols,
+                if uses_named_aggs:
+                    # If this is true, then we are dealing with agg with NamedAggregations.
+
+                    # When we have multiple columns with the same pandas label, we need to
+                    # generate dummy aggregations over all of the columns, as otherwise,
+                    # when we union the QueryCompilers, the duplicate columns will stack up
+                    # into one column. Example:
+                    #    A  A
+                    # 0  0  1
+                    # 1  2  3
+                    # If we call `df.agg(x=("A", "max"))` on the above DataFrame, we expect:
+                    #      A    A
+                    # x  2.0  NaN
+                    # x  NaN  3.0
+                    # but without the dummy aggregations, we get QC's that correspond to the following
+                    # frames:
+                    #      A
+                    # x  2.0
+                    # and
+                    #      A
+                    # x  3.0
+                    # so when we concatenate, we get a result that looks like this:
+                    #      A
+                    # x  2.0
+                    # x  3.0
+                    # which is wrong. Adding dummy aggregations means that our individual QCs will look like
+                    #      A    A
+                    # x  2.0  NaN
+                    # and
+                    #      A    A
+                    # x  NaN  3.0
+                    # so concatenation will give us the correct result.
+                    # We first check if it is the case that we are aggregating over multiple columns with the same pandas label
+                    has_col_with_duplicate_pandas_label = len(
+                        {id_pair.pandas_label for id_pair in column_to_agg_func.keys()}
+                    ) < len(column_to_agg_func.keys())
+
+                    def generate_single_agg_column_func_map(
+                        identifier_pair: PandasLabelToSnowflakeIdentifierPair,
+                        agg_func: AggFuncWithLabel,
+                    ) -> dict[PandasLabelToSnowflakeIdentifierPair, AggFuncInfo]:
+                        """
+                        Helper function to produce the agg func map for a single aggregation, including dummy aggregations.
+
+                        Notes:
+                        Adds dummy aggregations for all columns that will be aggregated over.
+                        """
+                        col_single_agg_func_map = {}
+                        for c in column_to_agg_func.keys():
+                            if c == identifier_pair:
+                                col_single_agg_func_map[c] = AggFuncInfo(
+                                    func=agg_func.func, is_dummy_agg=False
+                                )
+                            else:
+                                col_single_agg_func_map[c] = AggFuncInfo(
+                                    func="min", is_dummy_agg=True
+                                )
+                        return col_single_agg_func_map
+
+                    # We can't simply append the generated_qcs to single_agg_func_query_compilers, because
+                    # we have changed the order of the aggregations - see the comment in the function
+                    # extract_validate_and_try_convert_named_aggs_from_kwargs for an explanation of how
+                    # and why the order changes. Instead, we get a mapping of the name of the aggregation to the
+                    # QueryCompiler it produces, which we can then use to insert into single_agg_func_query_compilers
+                    # in the correct order, which we recieve from the frontend, so that when the concatenation occurs,
+                    # the final QueryCompiler is ordered correctly.
+                    index_label_to_generated_qcs: dict[
+                        Hashable, list["SnowflakeQueryCompiler"]
+                    ] = {}
+                    for (
+                        identifier_pair,
+                        agg_func_with_label,
+                    ) in column_to_agg_func.items():
+                        if not isinstance(agg_func_with_label, list):
+                            agg_func_with_label = [agg_func_with_label]
+                        for agg_func in agg_func_with_label:
+                            if not has_col_with_duplicate_pandas_label:
+                                agg_func_map = {
+                                    identifier_pair: AggFuncInfo(
+                                        agg_func.func, is_dummy_agg=False
+                                    )
+                                }
+                            else:
+                                agg_func_map = generate_single_agg_column_func_map(
+                                    identifier_pair, agg_func
+                                )
+
+                            new_qc = generate_agg_qc(
+                                agg_func_map,
+                                agg_func.pandas_label,
+                            )
+                            index_label_to_generated_qcs[
+                                agg_func.pandas_label
+                            ] = index_label_to_generated_qcs.get(
+                                agg_func.pandas_label, []
+                            ) + [
+                                new_qc
+                            ]
+                    correct_order_of_index_labels = list(kwargs.keys())
+                    for index_label in correct_order_of_index_labels:
+                        single_agg_func_query_compilers.extend(
+                            index_label_to_generated_qcs[index_label]
                         )
-                        for column in column_to_agg_func.keys()
-                    }
-                    single_agg_func_query_compilers.append(
-                        generate_agg_qc(
-                            col_single_agg_func_map, get_pandas_aggr_func_name(agg_func)
+                else:
+                    # get a map between each aggregation function and the columns needs to apply this aggregation function
+                    agg_func_to_col_map = get_agg_func_to_col_map(column_to_agg_func)
+
+                    # aggregation creates an index column with the aggregation function names as its values
+                    # For example: with following dataframe
+                    #       A   B   C
+                    #   0   1   2   3
+                    #   1   4   5   6
+                    #   2   7   8   9
+                    # after we call df.aggregate({"A": ["min"], "B": ["max"]}), the result is following
+                    #       A   B
+                    # min   1   NaN
+                    # max   NaN	8
+                    #
+                    # However, if all values in the agg_func dict are scalar strings/functions rather than lists,
+                    # then the result will instead be a Series:
+                    # >>> df.aggregate({"A": "min", "B": "max"})
+                    # 0    1
+                    # 1    8
+                    # dtype: int64
+                    for agg_func, cols in agg_func_to_col_map.items():
+                        col_single_agg_func_map = {
+                            column: AggFuncInfo(
+                                func=agg_func if column in cols else "min",
+                                is_dummy_agg=column not in cols,
+                            )
+                            for column in column_to_agg_func.keys()
+                        }
+                        single_agg_func_query_compilers.append(
+                            generate_agg_qc(
+                                col_single_agg_func_map,
+                                get_pandas_aggr_func_name(agg_func),
+                            )
                         )
-                    )
 
         assert single_agg_func_query_compilers, "no aggregation result"
         if len(single_agg_func_query_compilers) == 1:
@@ -8568,6 +8712,105 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
 
         return self
 
+    def dt_date(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("date")
+
+    def dt_time(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("time")
+
+    def dt_timetz(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("timetz")
+
+    def dt_year(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("year")
+
+    def dt_month(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("month")
+
+    def dt_day(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("day")
+
+    def dt_hour(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("hour")
+
+    def dt_minute(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("minute")
+
+    def dt_second(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("second")
+
+    def dt_microsecond(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("microsecond")
+
+    def dt_nanosecond(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("nanosecond")
+
+    def dt_dayofweek(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("dayofweek")
+
+    def dt_weekday(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("weekday")
+
+    def dt_dayofyear(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("dayofyear")
+
+    def dt_quarter(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("quarter")
+
+    def dt_is_month_start(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_month_start")
+
+    def dt_is_month_end(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_month_end")
+
+    def dt_is_quarter_start(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_quarter_start")
+
+    def dt_is_quarter_end(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_quarter_end")
+
+    def dt_is_year_start(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_year_start")
+
+    def dt_is_year_end(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_year_end")
+
+    def dt_is_leap_year(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("is_leap_year")
+
+    def dt_daysinmonth(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("daysinmonth")
+
+    def dt_days_in_month(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("days_in_month")
+
+    def dt_freq(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("freq")
+
+    def dt_seconds(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("seconds")
+
+    def dt_days(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("days")
+
+    def dt_microseconds(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("microseconds")
+
+    def dt_nanoseconds(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("nanoseconds")
+
+    def dt_components(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("components")
+
+    def dt_qyear(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("qyear")
+
+    def dt_start_time(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("start_time")
+
+    def dt_end_time(self) -> "SnowflakeQueryCompiler":
+        return self.dt_property("end_time")
+
     def dt_property(self, property_name: str) -> "SnowflakeQueryCompiler":
         """
         Extracts the specified date or time part from the timestamp.
@@ -11667,6 +11910,21 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
 
         return SnowflakeQueryCompiler(new_internal_frame)
 
+    def str_cat(
+        self,
+        others: ListLike,
+        sep: Optional[str] = None,
+        na_rep: Optional[str] = None,
+        join: Literal["left", "right", "outer", "inner"] = "left",
+    ) -> None:
+        ErrorMessage.method_not_implemented_error("cat", "Series.str")
+
+    def str_decode(self, encoding: str, errors: str) -> None:
+        ErrorMessage.method_not_implemented_error("decode", "Series.str")
+
+    def str_encode(self, encoding: str, errors: str) -> None:
+        ErrorMessage.method_not_implemented_error("encode", "Series.str")
+
     def str_startswith(
         self, pat: Union[str, tuple], na: object = None
     ) -> "SnowflakeQueryCompiler":
@@ -11705,6 +11963,37 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         """
         return self._str_startswith_endswith(pat, na, is_startswith=False)
 
+    def str_find(self, sub: str, start: int = 0, end: Optional[int] = None) -> None:
+        ErrorMessage.method_not_implemented_error("find", "Series.str")
+
+    def str_rfind(self, sub: str, start: int = 0, end: Optional[int] = None) -> None:
+        ErrorMessage.method_not_implemented_error("rfind", "Series.str")
+
+    def str_findall(self, pat: str, flags: int = 0) -> None:
+        ErrorMessage.method_not_implemented_error("findall", "Series.str")
+
+    def str_index(self, sub: str, start: int = 0, end: Optional[int] = None) -> None:
+        ErrorMessage.method_not_implemented_error("index", "Series.str")
+
+    def str_rindex(self, sub: str, start: int = 0, end: Optional[int] = None) -> None:
+        ErrorMessage.method_not_implemented_error("rindex", "Series.str")
+
+    def str_fullmatch(
+        self, pat: str, case: bool = True, flags: int = 0, na: object = None
+    ) -> None:
+        ErrorMessage.method_not_implemented_error("fullmatch", "Series.str")
+
+    def str_match(
+        self, pat: str, case: bool = True, flags: int = 0, na: object = None
+    ) -> None:
+        ErrorMessage.method_not_implemented_error("match", "Series.str")
+
+    def str_extract(self, pat: str, flags: int = 0, expand: bool = True) -> None:
+        ErrorMessage.method_not_implemented_error("extract", "Series.str")
+
+    def str_extractall(self, pat: str, flags: int = 0, expand: bool = True) -> None:
+        ErrorMessage.method_not_implemented_error("extractall", "Series.str")
+
     def str_capitalize(self) -> "SnowflakeQueryCompiler":
         """
         Capitalize the string
@@ -11723,6 +12012,12 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
         return SnowflakeQueryCompiler(new_internal_frame)
 
+    def str_isalnum(self) -> None:
+        ErrorMessage.method_not_implemented_error("isalnum", "Series.str")
+
+    def str_isalpha(self) -> None:
+        ErrorMessage.method_not_implemented_error("isalpha", "Series.str")
+
     def str_isdigit(self) -> "SnowflakeQueryCompiler":
         """
         Check whether all characters in each string are digits.
@@ -11737,6 +12032,9 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             )
         )
         return SnowflakeQueryCompiler(new_internal_frame)
+
+    def str_isspace(self) -> None:
+        ErrorMessage.method_not_implemented_error("isspace", "Series.str")
 
     def str_islower(self) -> "SnowflakeQueryCompiler":
         """
@@ -11797,6 +12095,12 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             )
         )
         return SnowflakeQueryCompiler(new_internal_frame)
+
+    def str_isnumeric(self) -> None:
+        ErrorMessage.method_not_implemented_error("isnumeric", "Series.str")
+
+    def str_isdecimal(self) -> None:
+        ErrorMessage.method_not_implemented_error("isdecimal", "Series.str")
 
     def str_lower(self) -> "SnowflakeQueryCompiler":
         """
@@ -11864,6 +12168,9 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         if flags & re.DOTALL:
             params = params + "s"
         return params
+
+    def str_center(self, width: int, fillchar: str = " ") -> None:
+        ErrorMessage.method_not_implemented_error("center", "Series.str")
 
     def str_contains(
         self,
@@ -11962,6 +12269,29 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
         return SnowflakeQueryCompiler(new_internal_frame)
 
+    def str_get(self, i: int) -> None:
+        ErrorMessage.method_not_implemented_error("get", "Series.str")
+
+    def str_get_dummies(self, sep: str) -> None:
+        ErrorMessage.method_not_implemented_error("get_dummies", "Series.str")
+
+    def str_join(self, sep: str) -> None:
+        ErrorMessage.method_not_implemented_error("join", "Series.str")
+
+    def str_pad(
+        self,
+        width: int,
+        side: Literal["left", "right", "both"] = "left",
+        fillchar: str = " ",
+    ) -> None:
+        ErrorMessage.method_not_implemented_error("pad", "Series.str")
+
+    def str_partition(self, sep: str = " ", expand: bool = True) -> None:
+        ErrorMessage.method_not_implemented_error("partition", "Series.str")
+
+    def str_rpartition(self, sep: str = " ", expand: bool = True) -> None:
+        ErrorMessage.method_not_implemented_error("rpartition", "Series.str")
+
     def str_len(self, **kwargs: Any) -> "SnowflakeQueryCompiler":
         """
         Compute the length of each element in the Series/Index
@@ -11981,6 +12311,15 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
                 lambda col: self._replace_non_str(col, length(col))
             )
         )
+
+    def str_ljust(self, width: int, fillchar: str = " ") -> None:
+        ErrorMessage.method_not_implemented_error("ljust", "Series.str")
+
+    def str_rjust(self, width: int, fillchar: str = " ") -> None:
+        ErrorMessage.method_not_implemented_error("rjust", "Series.str")
+
+    def str_normalize(self, form: Literal["NFC", "NFKC", "NFD", "NFKD"]) -> None:
+        ErrorMessage.method_not_implemented_error("normalize", "Series.str")
 
     def str_slice(
         self,
@@ -12117,6 +12456,14 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
         return SnowflakeQueryCompiler(new_internal_frame)
 
+    def str_slice_replace(
+        self,
+        start: Optional[int] = None,
+        stop: Optional[int] = None,
+        repl: Optional[Union[str, Callable]] = None,
+    ) -> None:
+        ErrorMessage.method_not_implemented_error("slice_replace", "Series.str")
+
     def str_split(
         self,
         pat: Optional[str] = None,
@@ -12162,6 +12509,8 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             )
         if pandas.isnull(regex):
             regex = False
+        if not pat and pat is not None:
+            raise ValueError("split() requires a non-empty pattern match.")
 
         if n is None:
             n = -1
@@ -12250,6 +12599,11 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             lambda col_name: output_col(col_name, pat, n)
         )
         return SnowflakeQueryCompiler(new_internal_frame)
+
+    def str_rsplit(
+        self, pat: Optional[str] = None, *, n: int = -1, expand: bool = False
+    ) -> None:
+        ErrorMessage.method_not_implemented_error("rsplit", "Series.str")
 
     def str_replace(
         self,
@@ -12376,6 +12730,15 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         )
         return SnowflakeQueryCompiler(new_internal_frame)
 
+    def str_repeat(self, repeats: int) -> None:
+        ErrorMessage.method_not_implemented_error("repeat", "Series.str")
+
+    def str_removeprefix(self, prefix: str) -> None:
+        ErrorMessage.method_not_implemented_error("removeprefix", "Series.str")
+
+    def str_removesuffix(self, prefix: str) -> None:
+        ErrorMessage.method_not_implemented_error("removesuffix", "Series.str")
+
     def str_strip(self, to_strip: Union[str, None] = None) -> "SnowflakeQueryCompiler":
         """
         Remove leading and trailing characters.
@@ -12406,6 +12769,24 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
             output_col
         )
         return SnowflakeQueryCompiler(new_internal_frame)
+
+    def str_lstrip(self, to_strip: Union[str, None] = None) -> None:
+        ErrorMessage.method_not_implemented_error("lstrip", "Series.str")
+
+    def str_rstrip(self, to_strip: Union[str, None] = None) -> None:
+        ErrorMessage.method_not_implemented_error("rstrip", "Series.str")
+
+    def str_swapcase(self) -> None:
+        ErrorMessage.method_not_implemented_error("swapcase", "Series.str")
+
+    def str_translate(self, table: dict) -> None:
+        ErrorMessage.method_not_implemented_error("translate", "Series.str")
+
+    def str_wrap(self, width: int, **kwargs: Any) -> None:
+        ErrorMessage.method_not_implemented_error("wrap", "Series.str")
+
+    def str_zfill(self, width: int) -> None:
+        ErrorMessage.method_not_implemented_error("zfill", "Series.str")
 
     def qcut(
         self,
@@ -12813,13 +13194,7 @@ class SnowflakeQueryCompiler(BaseQueryCompiler):
         return bins, SnowflakeQueryCompiler(ret_frame)
 
     def str_casefold(self) -> None:
-        """
-        Returns:
-            New query compiler with updated values.
-        """
-        ErrorMessage.not_implemented(
-            "Snowpark pandas doesn't yet support casefold method"
-        )
+        ErrorMessage.method_not_implemented_error("casefold", "Series.str")
 
     def dt_to_period(self, freq: Optional[str] = None) -> None:
         """