+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.18.1 (TBD)
+## 1.19.0 (TBD)
 
 ### Snowpark Python API Updates
 
@@ -10,8 +10,15 @@
 
 #### New Features
 
+#### Bug Fixes
+
+- Fixed a bug that causes output of GroupBy.aggregate's columns to be ordered incorrectly.
+
 #### Improvements
+
+- Added support for named aggregations in `DataFrame.aggregate` and `Series.aggregate` with `axis=0`.
 - `pd.read_csv` reads using the native pandas CSV parser, then uploads data to snowflake using parquet. This enables most of the parameters supported by `read_csv` including date parsing and numeric conversions. Uploading via parquet is roughly twice as fast as uploading via CSV.
+>>>>>>> 6b6663466d1af86cae17f1cf940c148d81d54e10
 
 ## 1.18.0 (2024-05-28)
 
@@ -27,6 +34,8 @@
 #### New Features
 
 - Added `DataFrame.cache_result` and `Series.cache_result` methods for users to persist DataFrames and Series to a temporary table lasting the duration of the session to improve latency of subsequent operations.
+
+#### Bug Fixes
 
 #### Improvements